Semi-functional horizontal approach
The application is working with a semi-horizontal microservice architecture approach.
A new microservice, web-server, has been added. The former module contains the AngularJS
app and, through its own simulated controllers and services, access the real ones, which
are dependant from the corresponding microservices.

Please, note that database access is not working properly in the user microservice and should be
looked into.

Another consideration would be to study the effects of the web-server microservice providing XML instead
of JSON.