Fixed hsqldb overriding MySQL
Eureka's hsqldb was overriding MySQL as a datasource and the user microservice was not able
to access its database. The problem was that Eureka's configuration .yml was overriding the
datasource. At the time, the datasource has been hard-coded when launching the microservice.
Ideally, such a hard-coded information should be read somehow from the .yml configuration file.