Handled server-side exceptions
Currently, exceptions in the backend are being handled irregardless of the cause of it. In production environments a user in the frontend shouldn't be provided with detailed information of what is happening in the backend.

Also, ResponseObject has been moved to a new package, org.sprouts.backend.pojo

A IntegerId class has been created so that Jackson can properly map Integers avoiding the workaround that was present in the UserController class. See http://stackoverflow.com/questions/19389723/can-not-deserialize-instance-of-java-lang-string-out-of-start-object-token for more information.