Fixed Jackson not mapping Integer
It turned out that Jackson was trying to deserialize the attribute 'id' of a User and treated it as a JSON_OBJECT instead of as a Integer instance.

Although creating an extra beam solved the issue, it looks like Jackson is capable of directly mapping the id attribute as the object itself. That is, passing '{"id":"5"}' can be automatically casted to a DomainEntity with id 5.